refactor(audio-bridge): moved participants type to common
