+++ jarust_plugins/src/audio_bridge/common.rs
@@ -12,3 +12,14 @@ pub enum Identifier {
     /// Unsigned Integer Identifier
     Uint(u64),
 }
+
+#[derive(PartialEq, Debug, Deserialize)]
+pub struct Participant {
+    pub id: Identifier,
+    pub display: Option<String>,
+    pub setup: bool,
+    pub muted: bool,
+    pub suspended: Option<bool>,
+    pub talking: Option<bool>,
+    pub spatial_position: Option<u64>,
+}
